\name{summary.PhenListAgeing}
\alias{summary.PhenListAgeing}

\title{
	Summary for a PhenListAgeing object
}
\description{
	This function provides a detailed summary of a PhenListAgeing object
}
\usage{
	\method{summary}{PhenListAgeing}(object, vars = NULL, ...)
}


\arguments{
	\item{object}{
		PhenListAgeing object 
	}
	\item{vars}{
		Variable(s) of interest
	}
	\item{\dots}{
		Optional parameters that can be passed to 'summarytools::dfSummary ()'
	}
}


\author{
	Hamed Haseli Mashhadi <hamedhm@ebi.ac.uk>
}

\seealso{
	\code{\link{PhenListAgeing}}
}

\examples{
	example(PhenListAgeing)
}
